Refactor to use new "memory buffer" objects, leading to significant speedup.
Eliminated a lot of the dynamic memory allocations that were even occurring for short messages (mostly reserving space in vectors) and eliminate the need to call GetRequiredSize type
functions for segments, bundles, etc. There is a BasicMemoryBuffer which is a quick way to add objects into contiguous memory, and a MemoryBuffer object as a concrete child
of BasicMemoryBuffer. MemoryBuffer has a known (non-type-template-parameter) stack size that it uses to until it eventually needs to dynamically allocate memory. Everything now formats to
buffer, instead of taking ranges of characters. Segments are now stored in a HybridVector object, which is a similar idea: it is a vector-like object where the first N elements are stored
in an internal, stack allocated buffer. I also found that storing the logger name as a string instead of a string view was slowing down logging a noticeable amount, since the name can be too
large for the string to use SBO.
All these optimizations speed up logging significantly, leading it to get ~5.5 million logs per second on my basic benchmark (compared with ~3 million before the refactor and ~6 million for
spdlog, equivalent test setup).
Updated profiling table in README.