Removed all RequiredSize functions
Now that formatting happens to buffers, it is completely unused. It is much faster to format log messages to one of the buffers than to use the RequiredSize functions and allocate a string.
It is also really hard to maintain, and not all the RequiredSize functions are even correct. Imagine actually implemented RequiredSize for a general formatted data type.